sha1_file: refactor into `find_pack_object_pos`
Looking up the offset in the packfile for a given SHA1 involves the
following:

	- Finding the position in the index for the given SHA1
	- Accessing the offset cache in the index for the found position

There are cases however where we'd like to find the position of a SHA1
in the index without looking up the packfile offset (e.g. when accessing
information that has been indexed based on index offsets).

This refactoring implements `find_pack_object_pos`, returning the
position in the index, and re-implements `find_pack_entry_one`(returning
the actual offset in the packfile) to use the new function.

Signed-off-by: Vicent Marti <tanoku@gmail.com>